feat: Add buffer registry refactor architectural plan (#375)
fix: active outputs and buffer sizes now propagate properly when the solver kernel is updated with new output types.
ai: agent pipeline updated to one taskmaster per task group, added a test agent.
fix: size-1 runs no longer break Cubie (closes #142)
fix: shared and local memory requirement names now consistent across package.

+++ src/cubie/outputhandling/summarymetrics/negative_peaks.py
@@ -95,7 +95,7 @@ class NegativePeaks(SummaryMetric):
             peak_counter = int32(buffer[2])
 
             if (
-                (current_index >= 2)
+                (current_index >= int32(2))
                 and (peak_counter < npeaks)
                 and (prev_prev != precision(0.0))
             ):
